+++ src/main/java/io/github/tr100000/researcher/api/ResearchRewardType.java
@@ -1,0 +1,7 @@
+package io.github.tr100000.researcher.api;
+
+import com.mojang.serialization.MapCodec;
+import net.minecraft.network.RegistryFriendlyByteBuf;
+import net.minecraft.network.codec.StreamCodec;
+
+public record ResearchRewardType<T extends ResearchReward>(MapCodec<T> codec, StreamCodec<RegistryFriendlyByteBuf, T> streamCodec) {}