+++ CHANGELOG.md
@@ -1,7 +1,14 @@
 ### Added
+- Added a research reward system
+  - Built-in reward types are `researcher:experience`, `researcher:fireworks`, and `researcher:loot`
+  - Mod authors can make and register their own reward types
 - Added research finished events (both client and server)
 
 ### Changed
 - Optimized png images
 - JSON field names now use snake case instead of camel case
 - Mess with data serialization
+- Made more breaking changes probably
+
+### Fixed
+- Bugs